add Pipfile.lock and restructure app
