+++ README.md
@@ -45,7 +45,7 @@ If all goes well, a `200` status is returned indicating the `alerts` have succes
 On connection to this endpoint, all of the `alerts` associated with the `object_id` (through the `object_alerts` join table) are pushed to the client. As additional alerts intended for the `object` with the `object_id`, they are also pushed to the client.
 
 # Setup
-A `Dockerfile` and `docker-compose.yml` were created but are admittedly a bit rough around the edges. This app pulls in environment variables from a `.env` in its `app` directory while you'll have to make; it should look like the following:
+A `Dockerfile` and `docker-compose.yml` were created but are admittedly a bit rough around the edges. This app pulls in environment variables from a `.env` in its root directory while you'll have to make; it should look like the following:
 
 ```
 DB_USERNAME=<YOUR_DB_USERNAME>
@@ -56,7 +56,7 @@ DB_NAME=<YOUR_DB_NAME>
 
 Indeed, the only environment variables we've got at the moment are the ones surrounding the database. Spool up a `PostgreSQL` database and paste the appropriate values into this `.env` file.
 
-Then, when you'd like to run this locally, navigate to the `app` directory and go ahead and run the following from the command line:
+Then, when you'd like to run this locally, navigate to the root directory and go ahead and run the following from the command line:
 
 ```bash
 uvicorn main:app
@@ -80,7 +80,7 @@ When hitting the `objects` endpoint, try sending up some objects with an invalid
 Have fun.
 
 # Tests
-Run `./tests.py` from the `app` directory. (You may need to run `chmod +x tests.py` if your system whines about not having sufficient privileges.)
+Run `./tests.py` from the directory. (You may need to run `chmod +x tests.py` if your system whines about not having sufficient privileges.)
 
 # TODO
 - [x] store credentials
