--- app/main.py
@@ -1,2 +1,0 @@
-from crud import *
-from websocket import *
