+++ app/models.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel
 from typing import List
+
+from pydantic import BaseModel
 
 class ObjectIn(BaseModel):
     email: str
